pyproject.toml
[tool.poetry]
name = "db-coursework"
version = "0.1.0"
description = ""
authors = ["Mikhail Belin <belinma@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"


[tool.poetry.group.lint.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
mypy = "^1.12.1"
types-psycopg2 = "^2.9.21.20241019"
types-requests = "^2.32.0.20241016"


[tool.poetry.group.dev.dependencies]
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = '.venv'

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.git
  )/
)
'''

[tool.isort]
line_length = 119